Clarify tests
